==> 01. Data Preprocessing.ipynb <==
## Importing necessary libs

import pandas as pd
import numpy as np
from pandas.io.json import json_normalize
import json

## Observing data format

df = pd.read_csv("data.csv").drop('Unnamed: 0', 1)

df.head()

### Parsing event data
Getting each event type

df["eventType"].unique()

df["eventType"]= df["eventType"].astype(str)

Getting the data structure for each event type

for uniqevent in df["eventType"].unique():
    print(uniqevent, ":", df[df["eventType"]==uniqevent]["event_data"].iloc[100])

Designing parsers for each data format

def extract_level_event_data(r):
    
    r = r.replace("=",":").replace('duration', '"duration"').replace("level", '"level"').replace("progress", '"progress"').replace("play_count", '"play_count"').replace("status", '"status"')
    r = json.loads(r)
    
    returndict = {
        "duration":0,
        "level":0,
        "progress":0,
        "play_count":0,
        "status":0,
    }
    
    returndict.update(r)
    
    return returndict["duration"], returndict["level"], returndict["progress"], returndict["play_count"], returndict["status"]

def extract_revenue_data(r):
    
    revenue = r.split(":")[1].split("'")[1]
    
    return revenue

Extracting event_data to seperate columns

df["duration"] = df["level"] = df["progress"] = df["play_count"] = df["status"] = df["revenue"] = 0

## extracting level data
df.loc[df["eventType"]=="level_event", ["duration", "level", "progress", "play_count", "status"]] =\
list(df['event_data'][df["eventType"]=="level_event"].apply(extract_level_event_data))

## extracting revenue data
df.loc[df["eventType"]!="level_event", "revenue"] =\
df['event_data'][df["eventType"]!="level_event"].apply(extract_revenue_data)

## object to float conversion
df["revenue"]= df["revenue"].astype(float)

Dropping json-like event_data

df.drop('event_data', 1, inplace=True)

df.head()

### Extracting meaningful event time feature

df["arrival_time"] = pd.to_datetime(df["arrival_time"])
df["event_time"] = pd.to_datetime(df["event_time"])

### Leaving only first 3 day events

days = (df["event_time"]-df["arrival_time"]).astype('timedelta64[D]')
df = df[days<=3]

Extracting customer lifetimes in minute domain.

df["event_delay"] = (df["event_time"]-df["arrival_time"]).astype('timedelta64[m]')
df.drop(['arrival_time', 'event_time'], 1, inplace=True)

### Reordering the columns

# Dropping the 'platform' since they are all = "ios"
df = df[["userID_anonymized", "model", "operating_system", "wifi_state", "session_no", "event_delay", "eventType", "duration", "level", "progress", "play_count", "status", "revenue", "totalrevenue"]]

df.head()

# Descriptive Analysis on Preprocessed Time Series Data

df.describe()

df.describe(include=['object', 'bool'])

df["model"].value_counts()[:10]

df["operating_system"].value_counts()[:10]

df["eventType"].value_counts()[:10]

Rewarded ads bring more revenue than interstitials, as expected.

for uniqevent in df["eventType"].unique():
    print("Revenues for", uniqevent, ">> mean:", np.round(df[df["eventType"]==uniqevent]["revenue"].mean(),3),\
         "|| std:", np.round(df[df["eventType"]==uniqevent]["revenue"].std(),3))

df.head()

## Saving pre-processed time-series data

df.to_csv('preprocessed_ts.csv', index=False)
