==> 02. Feature Extraction.ipynb <==
## Importing necessary libs

import pandas as pd
import numpy as np
from pandas.io.json import json_normalize
import json

## Reading time-series data

df = pd.read_csv("preprocessed_ts.csv")

df.head()

# Feature Extraction

I will extract these properties for each unique user:

- Total interstitial impression count
- Total banner impression count
- Total rewarded impression count
- Total level event count
- Last interstitial (delay_hour format)
- Last banner (delay_hour format)
- Last rewarded (delay_hour format)
- Last level event  (delay_hour format)

and mean and stds of the revenues of:
- Interstitials
- Banners
- Rewardeds

# Extracting means
means = df.groupby(['userID_anonymized', 'eventType'])['revenue'].mean().unstack(['eventType'], fill_value=0)\
.add_prefix('mean_revenue_').rename_axis(None, axis=1).reset_index().fillna(0).set_index("userID_anonymized")

# Extracting stds
stds = df.groupby(['userID_anonymized', 'eventType'])['revenue'].std().unstack(['eventType'], fill_value=0)\
.add_prefix('std_revenue_').rename_axis(None, axis=1).reset_index().fillna(0).set_index("userID_anonymized")

# Extracting total counts
totals = df.groupby(['userID_anonymized', 'eventType'])['revenue'].count().unstack(['eventType'], fill_value=0)\
.add_prefix('total_count_').rename_axis(None, axis=1).reset_index().fillna(0).set_index("userID_anonymized")

# Extracting last times
last_times = df.groupby(['userID_anonymized', 'eventType'])['event_delay'].max().unstack(['eventType'], fill_value=0)\
.add_prefix('last_time_').rename_axis(None, axis=1).reset_index().fillna(0).set_index("userID_anonymized")

### Concatenating extracted features

concatted_df = pd.concat([means, stds, totals, last_times], axis=1, join="inner").rename_axis(None, axis=1).reset_index()

### Merging with original data

concatted_df['model'] = df.groupby(['userID_anonymized'])['model'].min().values
concatted_df['operating_system'] = df.groupby(['userID_anonymized'])['operating_system'].min().values

# mean encoding for wi-fi
concatted_df['mean_wifi_state'] = df.groupby(['userID_anonymized'])['wifi_state'].mean().values

# to get last session number
concatted_df['last_session_no'] = df.groupby(['userID_anonymized'])['session_no'].max().values

# to get last played level
concatted_df['last_level'] = df.groupby(['userID_anonymized'])['level'].max().values

# to get lifetime-revenue
concatted_df['totalrevenue'] = df.groupby(['userID_anonymized'])['totalrevenue'].max().values

concatted_df.head()

## Saving feature-engineered data

concatted_df.to_csv('fe_data.csv', index=False)
