==> 03. Analysis.ipynb <==
## Importing necessary libs

import pandas as pd
from pandas_profiling import ProfileReport

## Reading feature-engineered data

df = pd.read_csv("fe_data.csv")

df.head()

df_profile = ProfileReport(df, title="DF Report")

df_profile.to_file("03 report.html")
